move internal metadata policy to control-plane and sc
