+++ crates/fluvio-controlplane-metadata/src/lib.rs
@@ -2,7 +2,6 @@ pub mod spu;
 pub mod topic;
 pub mod partition;
 pub mod spg;
-pub mod message;
 pub mod smartmodule;
 pub mod tableformat;
