+++ crates/fluvio-controlplane-metadata/src/partition/mod.rs
@@ -1,14 +1,10 @@
 mod spec;
 mod status;
-mod policy;
-mod replica;
-pub mod store;
 
 pub use self::spec::*;
 pub use self::status::*;
 pub use fluvio_protocol::record::ReplicaKey;
-pub use self::policy::*;
-pub use self::replica::*;
+
 
 #[cfg(feature = "k8")]
 mod k8;
