+++ crates/fluvio-controlplane/src/message/msg_type.rs
@@ -9,11 +9,12 @@ use std::fmt::Debug;
 use std::fmt::Display;
 use std::fmt;
 
+use fluvio_controlplane_metadata::core::MetadataItem;
+use fluvio_controlplane_metadata::core::Spec;
+use fluvio_controlplane_metadata::store::MetadataStoreObject;
+use fluvio_controlplane_metadata::store::actions::LSChange;
 use fluvio_protocol::{Encoder, Decoder};
 
-use crate::store::actions::*;
-use crate::core::*;
-use crate::store::*;
 
 #[derive(Decoder, Default, Encoder, Debug, Eq, PartialEq, Clone)]
 pub enum MsgType {
