+++ crates/fluvio-controlplane/src/requests/mod.rs
@@ -1,9 +1,2 @@
-pub mod update_spu;
-pub mod update_replica;
-pub mod register_spu;
-pub mod update_lrs;
-pub mod remove;
-pub mod update_smartmodule;
-
 mod request;
-pub use self::request::ControlPlaneRequest;
+pub use self::request::*;
