+++ crates/fluvio-controlplane/src/sc_api/mod.rs
@@ -1,0 +1,4 @@
+pub mod api;
+pub mod register_spu;
+pub mod remove;
+pub mod update_lrs;
