+++ crates/fluvio-controlplane/src/spu_api/mod.rs
@@ -1,0 +1,4 @@
+pub mod api;
+pub mod update_replica;
+pub mod update_smartmodule;
+pub mod update_spu;
