+++ crates/fluvio-controlplane/src/spu_api/update_replica.rs
@@ -3,11 +3,13 @@
 use fluvio_protocol::api::Request;
 use fluvio_protocol::Decoder;
 use fluvio_protocol::Encoder;
-use fluvio_controlplane_metadata::partition::Replica;
 
-use crate::InternalSpuApi;
 
-use super::ControlPlaneRequest;
+use crate::replica::Replica;
+use crate::requests::ControlPlaneRequest;
+
+use super::api::InternalSpuApi;
+
 
 pub type UpdateReplicaRequest = ControlPlaneRequest<Replica>;
