+++ crates/fluvio-sc/src/stores/spu/mod.rs
@@ -1,3 +1,6 @@
+mod store;
+
+pub use store::*;
 pub use fluvio_controlplane_metadata::spu::store::*;
 pub use fluvio_controlplane_metadata::spu::*;
 pub use fluvio_controlplane_metadata::store::k8::K8MetaItem;
@@ -5,6 +8,7 @@ pub use health_check::*;
 
 pub type SpuAdminMd = SpuMetadata<K8MetaItem>;
 pub type SpuAdminStore = SpuLocalStore<K8MetaItem>;
+
 
 // check if given range is conflict with any of the range
 pub async fn is_conflict(
